Controls4U. Single threaded compiling (non MT) is now permitted, losing StaticClock auto refresh and Meter handle movement
git-svn-id: svn://ultimatepp.org/upp/trunk@2448 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ bazaar/Controls4U/Controls4U.h
@@ -315,7 +315,7 @@ protected:
 	bool clockWise;
 	bool number;
 	int colorType;
-	int speed;
+	double speed;
 	int sensibility;
 	volatile Atomic running, kill;
 	
@@ -340,7 +340,7 @@ public:
 	Meter& ClockWise(bool v)		{clockWise = v; Refresh(); return *this;}
 	Meter& SetNumber(bool v)		{number = v; Refresh(); return *this;}
 	Meter& SetColorType(int c)		{colorType = c; Refresh(); return *this;}
-	Meter& SetSpeed(int s)			{speed = s; Refresh(); return *this;} 
+	Meter& SetSpeed(double s)		{speed = s; Refresh(); return *this;} 
 	Meter& SetSensibility(int s)	{sensibility = s; Refresh(); return *this;}
 
 	void SetData(const Value& v);
